levels58/requirements.txt
Flask==0.10.1
gunicorn==19.4.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysql==0.0.1
MySQL-python==1.2.5
Werkzeug==0.11.5
